pyproject.toml
[tool.poetry]
name = "llavacap"
version = "0.1.0"
description = "LLM captioning based on LLaVa model"
authors = ["Anton Nikonov <ant.nik.nik@mail.ru>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "4.37.2"
bitsandbytes = "0.41.3"
accelerate = "0.25.0"
scipy = "1.14.0"
pillow = "10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
